Interpolation format for tourr

Generates the interpolation in the form needed for tourr, modelled on geodesic_path()

Usage

givens_path(current, target, frozen = NULL, ...)

Arguments

current

starting frame

target

target frame

frozen

indicator whether some dimensions fixed

...

arguments sent to later functions

Value

interpolate

A function with single parameter in [0, 1] that returns an interpolated frame between the current and future frames. 0 gives the current plane, 1 gives the new target frame in plane of current frame.

dist

The distance, in radians, between the current and target frames.

Fa

The current frame.

Fz

The new target frame.

tau

The principle angles between the current and target frames.

Ga

The current plane.

Gz

The target plane.