woylier: Alternative Tour Frame Interpolation Method

This method generates a tour path by interpolating between d-D frames in p-D using Givens rotations. The algorithm arises from the problem of zeroing elements of a matrix. This interpolation method is useful for showing specific d-D frames in the tour, as opposed to d-D planes, as done by the geodesic interpolation. It is useful for projection pursuit indexes which are not s invariant. See more details in Buj, Cook, Asimov and Hurley (2005) doi:10.1016/S0169-7161(04)24014-7 and Batsaikhan, Cook and Laa (2023) doi:10.48550/arXiv.2311.08181 .

See also

Author

Maintainer: Dianne Cook dicook@monash.edu (ORCID)

Authors:

Dianne Cook dicook@monash.edu (ORCID)

Zola Batsaikan zoljargal11@gmail.com (ORCID)

Ursula Laa ursula.laa@boku.ac.at (ORCID)